Hierarchical Category Inserting implemented
Hierarchical Category Inserting (which means that the ParentId is set right) is done automatically,
because the structure of the imported csv-Entry of Categories allowes to get the master category first and
then the sub category. Thus, the parents id is known beforehand.
So adding the column ParentCategoryId to the Categorys table was sufficient.

For the transaction part of the YNAB import only the split transaction are left to be investigated.
Afterwards the budgeting follows. Yeah.