Restructured the MainWindow into another root abstraction level.
Now the MainWindow is responsible mainly for the flyouts. Opening, Importing and Creating a new budget plan. And the title commands.
The Content is delegated to the FilledMainWindow, which for the moment holds only the Account Tabs.
Also included an empty MainWindow option if the database path does not lead to an existing file.
Refactored the redundant references to libraries out of most of the classes. I will try to do these in separate commits next time.