Instead of switching the TitContentViewModel now the existing one gets reset.
The UserControl corresponding to the TitContentViewModels is now in a DataTemplate. These are cached in WPF and depend on the DataType. This means only one DataTemplate instance per Type. Therefore the easier way is to just update the existing TitContentViewModel.

IoC container now only injects from App.xaml.cs, because the "SingletonScope" in Ninject is NOT a singleton, but apparently copies. The problem is the private variables will not be copied, but get their default values. This is a crucial flaw, because the SqLiteBffOrm use a private lock flag, which can cause serious misbehavior.

Again a big commit. Sorry again. I'll improve that eventually.