Reworked the Culture-issue. This is a big commit.
Introducing the BffEnvironment, which will hold the dependencies for the whole project, which are used in many places. Starting with the CultureProvider, which holds the Cultures for Language, Currency and Date. Now every use of the Cultures will be tightly coupled to the BffEnvironment, but this is okay for now.
I tried an approach with the IoC-container of Ninject, but this had several problems:
- The "SingletonScope" does not imply real singletons, so private variables get the default values and not the current. On different place actually different instances of the dependent class are worked with
- Pretty difficult to inject the dependencies into Converters for example. MultiBindings would be one solution, but ...
- It produces much boilerplate code.