Restructured and refactored all of the native model classes.
Now they are cleaner, use the advantages of OOP and have comments on almost all public members.
The models now have both a save to use constructor for OR-mapping and a constructor, which is almost completely optional for property initialization.
The constructor have to be preferred as the won't trigger any updates on construction because of properties.

Even though this commit is big, I am proud of it, because it has only one topic and is consequent.