Now the Payee- and Category-Editing-Forms do stretch to full length
A StackPanel cannot stretch its containing controls to full length in its Orientation direction.
The DockPanel with its LastChildFill property does the trick. With that option the last child of the DockPanel fills the remaining space. I like it, because with that option you can control, which element exactly should be stretched.