Alternative way of handling the OR-mapping of tits.
Now in account-specific and account-unspecific the tits have the same object base and don't need to be refreshe manually on each change. Inserting and deleting is handle by the ORM.
The filtering is excluded for now.
This solution is pretty stress free, but I don't think that it scales well. For the firt 500 rows it should be okay, though. Until then another solution has to be found.
Like Data Virtualization.