At this branch the AlphaChiTech Virtualization implementation is tested. For now AllAccounts works as far as to read in the data correctly.
This one is out of the box well working.
- The scrolling speed is amazing
- Pretty simple to integrate. This commit is somewhat dirty, but it can be worked out more elegant.
- It supports editing already, but …

One contra is that calling the first editing form is taking alot of time. All following calls are instantly processed. I have to think about this trade-off or find a solution for it.

Another point is, that this first try is sync. The framework supports async access, so it maybe worth checking this out.