Uni-sizing the normal and editing cells (MinHeight = 30) and centering all CellContent
This workaround solves the problem the the virtualization break in the VirtualizingStackPanel.
It seems that if a row changes it height, then the VirtualizingStackPanel fetches all items.

The problem will still occur with Parent-Tits, because it uses the RowDetails, but these will probably be less frequently used. Later, this can be solved by a PopUp or a flyout.

Still one problem to solve before this data virtualization solution is fit for the develop branch:
To refresh the contents Clear() is used so far, this also breaks the virtualization. Because the collection sends a Reset-Event to the DataGrid, which then fetches all items again.