Bugfix: Summary Account refreshes the starting and current balance on account addition.
The problem was with the Messenger class it uses the MessengerKey to store Actions for certain recipients and contexts. The MessengerKey did not consider the message type yet, so only one message type (the first) could be registered with the same context.
I fixed that by adding the type of the message to the MessengerKey.

Also rename refactor: AllAccountMessage to SummaryAccountMessage

Signed-off-by: Dima <dieter.enns@gmail.com>